Rename `tag_column_aliases.py`.
