+++ metricflow/sql/optimizer/column_pruning/required_column_aliases.py
@@ -9,7 +9,7 @@ from metricflow_semantics.sql.sql_exprs import SqlExpressionTreeLineage
 from typing_extensions import override
 
 from metricflow.sql.optimizer.column_pruning.cte_alias_to_cte_node_mapping import SqlCteAliasMappingLookup
-from metricflow.sql.optimizer.column_pruning.tag_column_aliases import NodeToColumnAliasMapping
+from metricflow.sql.optimizer.column_pruning.node_to_column_alias_maping import NodeToColumnAliasMapping
 from metricflow.sql.sql_ctas_node import SqlCreateTableAsNode
 from metricflow.sql.sql_cte_node import SqlCteAliasMapping, SqlCteNode
 from metricflow.sql.sql_plan import (
